[test] Rename to SimpleTableTestBase and refactor overwrite tests
